# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: resource not found
        type: string
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.Group:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      id:
        example: 1
        type: integer
      is_default:
        example: false
        type: boolean
      name:
        example: Administrators
        type: string
      organization:
        $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Organization'
      organization_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      users:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.User'
        type: array
    required:
    - name
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  github_com_eenemeene_kitamanager-go_internal_models.LoginResponse:
    properties:
      expires_in:
        example: 3600
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  github_com_eenemeene_kitamanager-go_internal_models.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  github_com_eenemeene_kitamanager-go_internal_models.Organization:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Group'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: Acme Corp
        type: string
      state:
        example: berlin
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    required:
    - name
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.User:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      email:
        example: john@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Group'
        type: array
      id:
        example: 1
        type: integer
      is_superadmin:
        example: false
        type: boolean
      last_login:
        example: "2024-01-15T10:30:00Z"
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    required:
    - email
    - name
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.UserResponse:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      email:
        example: john@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Group'
        type: array
      id:
        example: 1
        type: integer
      is_superadmin:
        example: false
        type: boolean
      last_login:
        example: "2024-01-15T10:30:00Z"
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@kitamanager.example.com
    name: API Support
  description: REST API for managing Users, Groups, and Organizations with RBAC support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: KitaManager API
  version: "1.0"
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /api/v1/logout:
    post:
      description: Clear authentication cookies to log out the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      summary: Logout user
      tags:
      - auth
  /api/v1/me:
    get:
      description: Returns the currently authenticated user based on the JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
...
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for new access and refresh tokens
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"